Add agreement compoent
